Adds worker remove functionality

+++ lib/run_rabbit_run/utils/signals.rb
@@ -18,44 +18,28 @@ module RRR
           return true
         rescue Errno::ESRCH
         rescue Exception => e
-          RRR.logger.error "#{e}, #{e.backtrace.join("\n")}"
+          RRR.logger.error e.message
         end
 
         return false
       end
 
-      def kill_signal name, pid
-        send_signal name, pid, RRR::Utils::Signals::KILL
+      def kill_signal pid
+        send_signal pid, RRR::Utils::Signals::KILL
       end
 
-      def stop_signal name, pid
-        send_signal name, pid, RRR::Utils::Signals::QUIT
+      def stop_signal pid
+        send_signal pid, RRR::Utils::Signals::QUIT
       end
 
-      def reload_signal name, pid
-        send_signal name, pid, RRR::Utils::Signals::RELOAD
+      def reload_signal pid
+        send_signal pid, RRR::Utils::Signals::RELOAD
       end
 
     private
 
-      def send_signal name, pid, signal
-        if running? pid
-          RRR.logger.info "[#{name}] send #{signal_name(signal)} signal to process"
-          Process.kill(signal, pid)
-        else
-          RRR.logger.debug "[#{name}] is not running"
-        end
-      end
-
-      def signal_name code
-        case code
-        when 'QUIT'
-          'exit'
-        when 'USR1'
-          'reload'
-        when 'KILL'
-          'kill'
-        end
+      def send_signal pid, signal
+        Process.kill(signal, pid) if running? pid
       end
     end
   end
